# Preset capacity for ApptibuteMap and StringMap to improve performance