--- a/translator/trace/jaeger/jaegerproto_to_traces.go
+++ b/translator/trace/jaeger/jaegerproto_to_traces.go
@@ -101,7 +101,10 @@ func jProcessToInternalResource(process *model.Process, dest pdata.Resource) {
 
 	attrs := dest.Attributes()
 	if serviceName != "" {
+		attrs.InitEmptyWithCapacity(len(tags) + 1)
 		attrs.UpsertString(conventions.AttributeServiceName, serviceName)
+	} else {
+		attrs.InitEmptyWithCapacity(len(tags))
 	}
 	jTagsToInternalAttributes(tags, attrs)
 
@@ -163,6 +166,7 @@ func jSpanToInternal(span *model.Span, dest pdata.Span) {
 	}
 
 	attrs := dest.Attributes()
+	attrs.InitEmptyWithCapacity(len(span.Tags))
 	jTagsToInternalAttributes(span.Tags, attrs)
 	setInternalSpanStatus(attrs, dest.Status())
 	if spanKindAttr, ok := attrs.Get(tracetranslator.TagSpanKind); ok {
@@ -298,6 +302,7 @@ func jLogsToSpanEvents(logs []model.Log, dest pdata.SpanEventSlice) {
 		}
 
 		attrs := event.Attributes()
+		attrs.InitEmptyWithCapacity(len(log.Fields))
 		jTagsToInternalAttributes(log.Fields, attrs)
 		if name, ok := attrs.Get("message"); ok {
 			event.SetName(name.StringVal())