--- a/translator/trace/jaeger/jaegerthrift_to_traces.go
+++ b/translator/trace/jaeger/jaegerthrift_to_traces.go
@@ -66,7 +66,10 @@ func jThriftProcessToInternalResource(process *jaeger.Process, dest pdata.Resour
 
 	attrs := dest.Attributes()
 	if serviceName != "" {
+		attrs.InitEmptyWithCapacity(len(tags) + 1)
 		attrs.UpsertString(conventions.AttributeServiceName, serviceName)
+	} else {
+		attrs.InitEmptyWithCapacity(len(tags))
 	}
 	jThriftTagsToInternalAttributes(tags, attrs)
 
@@ -108,6 +111,7 @@ func jThriftSpanToInternal(span *jaeger.Span, dest pdata.Span) {
 	}
 
 	attrs := dest.Attributes()
+	attrs.InitEmptyWithCapacity(len(span.Tags))
 	jThriftTagsToInternalAttributes(span.Tags, attrs)
 	setInternalSpanStatus(attrs, dest.Status())
 	if spanKindAttr, ok := attrs.Get(tracetranslator.TagSpanKind); ok {
@@ -159,6 +163,7 @@ func jThriftLogsToSpanEvents(logs []*jaeger.Log, dest pdata.SpanEventSlice) {
 		}
 
 		attrs := event.Attributes()
+		attrs.InitEmptyWithCapacity(len(log.Fields))
 		jThriftTagsToInternalAttributes(log.Fields, attrs)
 		if name, ok := attrs.Get("message"); ok {
 			event.SetName(name.StringVal())